Remove duplicated frame time rendering
The frame time is rendered in the same place by both ui::ui() as well as
run_app(). The former renders the raw rendering time, the latter
includes the time for any performed actions, and only that is actually
visible.

So we can remove the duplicate code, and show only the action+render
time. As we want that to happen in ui::ui(), we need to make the start
time part of the app struct.